feat: refactor pilot logging to include a business logic layer
