+++ diracx-db/src/diracx/db/os/pilot_logs.py
@@ -19,3 +19,17 @@ class PilotLogsDB(BaseOSDB):
         # TODO decide how to define the index name
         # use pilot ID
         return f"{self.index_prefix}_{doc_id // 1e6:.0f}"
+
+
+async def search_message(db: PilotLogsDB, search_params: list[dict]):
+
+    return await db.search(
+        ["Message"],
+        search_params,
+        [{"parameter": "LineNumber", "direction": "asc"}],
+    )
+
+
+async def bulk_insert(db: PilotLogsDB, docs: list[dict], pilot_id: int):
+
+    await db.bulk_insert(db.index_name(pilot_id), docs)
